tools/python/requirements.txt
boto3==1.26.155
docopt==0.6.2
pytest==7.3.2
Jinja2==3.1.2
mergedeep==1.3.4
pyyaml
wcmatch==8.4.1
openai==0.27.8
azure-identity==1.13.0
azure.mgmt.network==23.1.0